Handle path lengths > 200 characters
We now use the "data" field of the list-box% to hold the actual file
path, and use a shortened version of the path as label.

The width of the box is used to estimate the length where shortening
must be applied; the shortening itself simply removes some characters
from the middle of the string.